Play song by id, not by position

+++ src/rust/entities.rs
@@ -28,7 +28,7 @@ pub struct QSong {
 pub enum MPSCCommand {
     Next,
     Previous,
-    PlaySong(usize),
+    PlaySong(u64),
     PlayToggle,
     UpdateDb,
     GetPlaylists(SongField),
